server/requirements.txt
alabaster==0.7.16
amqp==5.2.0
anytree==2.12.1
asgiref==3.8.1
attrs==23.2.0
autopage==0.5.2
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.0
bullet==2.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cliff==4.7.0
cmd2==2.4.3
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
dj-config-url==0.1.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.0.0
docutils==0.21.2
drf-yasg==1.21.7
filelock==3.14.0
gnupg==2.3.1
humanfriendly==10.0
idna==3.7
imagesize==1.4.1
inflection==0.5.1
ipwhois==1.2.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
logging==0.4.9.6
lxml==5.2.2
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.0
pbr==6.0.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prettytable==3.10.0
prompt-toolkit==3.0.43
property-manager==3.0
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyperclip==1.8.2
pyproject-api==1.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
setuptools==69.5.1
setuptools-scm==8.1.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.0
stevedore==5.2.0
tox==4.15.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
verboselogs==1.7
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
xkcdpass==1.19.9
yamlreader==3.0.4
